use unified reference_clips instead of old worm specific one
